refactor : entity 디렉터리, response message enum 클래스 위치 변경
data 디렉터리 하위로 이동!

+ 관련 변경 사항 커밋

issue key : SB2-7
