+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/data/NicknameDto.java
@@ -1,6 +1,5 @@
 package kim.zhyun.serveruser.data;
 
-import kim.zhyun.serveruser.entity.SessionUser;
 import lombok.*;
 
 import java.util.Objects;
