+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/data/entity/User.java
@@ -1,4 +1,4 @@
-package kim.zhyun.serveruser.entity;
+package kim.zhyun.serveruser.data.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
