+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/interceptor/ConnectionInterceptor.java
@@ -3,7 +3,7 @@ package kim.zhyun.serveruser.interceptor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.data.entity.SessionUser;
 import kim.zhyun.serveruser.service.SessionUserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
