+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/repository/SessionUserRepository.java
@@ -1,6 +1,6 @@
 package kim.zhyun.serveruser.repository;
 
-import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.data.entity.SessionUser;
 import org.springframework.data.repository.CrudRepository;
 
 public interface SessionUserRepository extends CrudRepository<SessionUser, String> {
