+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package kim.zhyun.serveruser.repository;
 
-import kim.zhyun.serveruser.entity.User;
+import kim.zhyun.serveruser.data.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
