+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/service/impl/NicknameReserveServiceImpl.java
@@ -1,7 +1,7 @@
 package kim.zhyun.serveruser.service.impl;
 
 import kim.zhyun.serveruser.data.NicknameDto;
-import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.data.entity.SessionUser;
 import kim.zhyun.serveruser.repository.SessionUserRepository;
 import kim.zhyun.serveruser.service.NicknameReserveService;
 import lombok.Getter;
