+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/service/impl/SignUpServiceImpl.java
@@ -1,7 +1,7 @@
 package kim.zhyun.serveruser.service.impl;
 
 import kim.zhyun.serveruser.data.NicknameDto;
-import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.data.entity.SessionUser;
 import kim.zhyun.serveruser.repository.UserRepository;
 import kim.zhyun.serveruser.service.NicknameReserveService;
 import kim.zhyun.serveruser.service.SessionUserService;
