+++ simpleBoard02/server-b-user/src/test/java/kim/zhyun/serveruser/controller/CheckControllerTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static kim.zhyun.serveruser.data.type.ResponseMessage.*;
+import static kim.zhyun.serveruser.data.message.ResponseMessage.*;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
