+++ simpleBoard02/server-b-user/src/test/java/kim/zhyun/serveruser/interceptor/ConnectionInterceptorTest.java
@@ -2,7 +2,7 @@ package kim.zhyun.serveruser.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.data.entity.SessionUser;
 import kim.zhyun.serveruser.service.SessionUserService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
