+++ simpleBoard02/server-b-user/src/test/java/kim/zhyun/serveruser/repository/NicknameStorageTest.java
@@ -1,7 +1,7 @@
 package kim.zhyun.serveruser.repository;
 
 import kim.zhyun.serveruser.data.NicknameDto;
-import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.data.entity.SessionUser;
 import kim.zhyun.serveruser.repository.container.RedisTestContainer;
 import kim.zhyun.serveruser.service.NicknameReserveService;
 import kim.zhyun.serveruser.service.SessionUserService;
